Add player-specific rate limiting command (ifnxplslp).
This is based off of a patch qt wrote for dgamelaunch but was never
implemented until now. Similar to the 'ifnxsleep' command that was
developed previously, this command allows for specifying a player
instead of a blanket limiter for everyone that plays a specific game.
The command will delay game execution for X amount of seconds if the
player name specified matches the user logged in AND if no save file
exists for that player.