Add IP-based rate limiting command (ifnxipslp).
Much like the newly added ifnxplslp command, ifnxipslp checks against an
IP address as the source instead of a specific player name. Example:
player Bob logs in and fires up a startscumming script, and then has his
name added to a ifnxplslp command line. Bob decides to cirumvent that
delay by creating a new account and logs in as Bob2. Admin notices that
Bob and Bob2 come from the same IP address, and utilizes the ifnxipslp
command line using Bob's IP address. Now, no matter what accounts Bob
makes to try and get around the delay, the delay will be active as long
as Bob tries to connect from that IP address.

Could Bob fire up a private VPN service and try to get around the delay
by connecting from a different IP? Sure, but it may be too much work and
hassle for Bob to go through the trouble of doing this.